+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 [1.1.0] 08.04.2026
 
+- [FAM-7] Transactions screen
 - [FAM-6] Wallets in dashboard
 - [FAM-5] Dashboard